feat: interview que. on counter
after setcounter(counter + 1) many times what is the output?
here the concept of useState role comes and it send code in batches so it only increment by one.
And other side if we use with collback which add previous value to next if we run that many times then it can increase value with addind previous value.